+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ryanatkn/gro
 
+## 0.189.3
+
+### Patch Changes
+
+- improve `gro format` to take rest and forwarded args ([c9de33f](https://github.com/ryanatkn/gro/commit/c9de33f))
+
+## 0.189.2
+
+### Patch Changes
+
+- add `--build` and `--force_build` flags to `gro check` ([2557e50](https://github.com/ryanatkn/gro/commit/2557e50))
+
 ## 0.189.1
 
 ### Patch Changes